+++ server/src/test/java/org/opensearch/index/translog/RemoteFSTranslogTests.java
@@ -107,6 +107,7 @@ public class RemoteFSTranslogTests extends OpenSearchTestCase {
     protected Path translogDir;
     // A default primary term is used by translog instances created in this test.
     private final AtomicLong primaryTerm = new AtomicLong();
+    private final AtomicBoolean primaryMode = new AtomicBoolean(true);
     private final AtomicReference<LongConsumer> persistedSeqNoConsumer = new AtomicReference<>();
     private ThreadPool threadPool;
 
@@ -170,7 +171,8 @@ public class RemoteFSTranslogTests extends OpenSearchTestCase {
             primaryTerm::get,
             getPersistedSeqNoConsumer(),
             repository,
-            threadPool.executor(ThreadPool.Names.TRANSLOG_TRANSFER)
+            threadPool.executor(ThreadPool.Names.TRANSLOG_TRANSFER),
+            primaryMode::get
         );
 
     }
@@ -250,6 +252,33 @@ public class RemoteFSTranslogTests extends OpenSearchTestCase {
         translog.ensureSynced(loc);
         list.add(op);
         return loc;
+    }
+
+    public void testUploadWithPrimaryModeFalse() {
+        // Test setup
+        primaryMode.set(false);
+
+        // Validate
+        assertTrue(translog.syncNeeded());
+        assertFalse(primaryMode.get());
+        try {
+            translog.sync();
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+        assertTrue(translog.syncNeeded());
+    }
+
+    public void testUploadWithPrimaryModeTrue() {
+        // Validate
+        assertTrue(translog.syncNeeded());
+        assertTrue(primaryMode.get());
+        try {
+            translog.sync();
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+        assertFalse(translog.syncNeeded());
     }
 
     public void testSimpleOperations() throws IOException {
@@ -1120,7 +1149,8 @@ public class RemoteFSTranslogTests extends OpenSearchTestCase {
                 primaryTerm::get,
                 persistedSeqNos::add,
                 repository,
-                threadPool.executor(ThreadPool.Names.TRANSLOG_TRANSFER)
+                threadPool.executor(ThreadPool.Names.TRANSLOG_TRANSFER),
+                () -> Boolean.TRUE
             ) {
                 @Override
                 ChannelFactory getChannelFactory() {